refactor(direct-control): prune session facade type imports
Refs: openspec/changes/civ7-support-direct-control-modularization

Repoints remaining internal type-only imports for command results, direct-control options, state selections, sessions, runtime probes, snapshot/health results, and operation families at their owner modules instead of the public facade. Public facade re-exports remain stable for package consumers, and runtime behavior is unchanged.

This keeps procedure schemas, operation/proof telemetry, AI ingestion, semantic CLI projection, hotseat runtime proof, Effect/Bun composition, Effect/oRPC procedure cores, and Task 2.9.4 matrix-row acceptance pending.

Validation:
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test -- play-notification-view.test.ts production-choice.test.ts diplomacy-response.test.ts narrative-choice.test.ts unit-operation.test.ts setup-and-lifecycle.test.ts restart-lifecycle.test.ts runtime-and-catalog.test.ts public-api.test.ts
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- rg ownership scan for internal facade back-imports under packages/civ7-direct-control/src
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- git diff --cached --check

Local type/source relocation proof only; no runtime, live-game, AI-on-hotseat, telemetry, semantic CLI, Effect/oRPC, or procedure-core proof claimed.
