+++ packages/civ7-direct-control/src/play/operations/diplomacy-postconditions.ts
@@ -2,7 +2,7 @@ import type { Civ7DiplomacyResponseInput } from "./diplomacy-request.js";
 import type { Civ7OperationValidationResult } from "./types.js";
 import type { Civ7ComponentId } from "../../civ7-component-id.js";
 import type { Civ7RuntimeProbe } from "../../runtime/probe.js";
-import type { Civ7DirectControlOptions } from "../../index.js";
+import type { Civ7DirectControlOptions } from "../../session/types.js";
 import type {
   Civ7PlayNotificationSummary,
   Civ7PlayNotificationViewResult,
